downloader: refactor downloadRepository function
