+++ downloader/git.go
@@ -3,7 +3,12 @@ package downloader
 import (
 	"context"
 	"fmt"
+	"io"
+	"os"
+	"path/filepath"
 
+	"github.com/src-d/go-borges"
+	"github.com/src-d/go-borges/siva"
 	"gopkg.in/src-d/go-billy.v4"
 	"gopkg.in/src-d/go-billy.v4/util"
 	"gopkg.in/src-d/go-errors.v1"
@@ -29,7 +34,9 @@ const (
 	fetchRefSpecStr = "+refs/*:refs/remotes/%s/*"
 )
 
-func cloneRepo(
+// CloneRepository clones a git repository from the given endpoint into the
+// billy.Filesystem. A remote with the id is created for that.
+func CloneRepository(
 	ctx context.Context,
 	fs billy.Filesystem,
 	path, endpoint, id, token string,
@@ -106,6 +113,29 @@ func createRemote(r *git.Repository, id, endpoint string) (*git.Remote, error) {
 	return r.Remote(id)
 }
 
+// RootCommit traverse the commit history for the given remote following the
+// first parent of each commit. The root commit found (commit with no parents)
+// is returned.
+func RootCommit(
+	repo *git.Repository,
+	remote string,
+) (*object.Commit, error) {
+	start, err := headCommit(repo, remote)
+	if err != nil {
+		return nil, err
+	}
+
+	current := start
+	for len(current.ParentHashes) > 0 {
+		current, err = current.Parent(0)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	return current, nil
+}
+
 func headCommit(repo *git.Repository, id string) (*object.Commit, error) {
 	ref, err := repo.Reference(
 		plumbing.NewRemoteHEADReferenceName(id),
@@ -138,24 +168,194 @@ func resolveCommit(
 	}
 }
 
-// rootCommitHash traverse the commit history from the given start commit
-// following the first parent of each commit. The root commit found (commit
-// with no parents) is returned.
-func rootCommit(
-	repo *git.Repository,
-	start *object.Commit,
-) (*object.Commit, error) {
-	var (
-		current = start
-		err     error
-	)
+// PrepareRepository returns a borges.Repository ready to fetch changes.
+// It creates a rooted repository copying the cloned repository in tmp to
+// the siva file the library uses at the location with the given location ID,
+// creating this location if not exists.
+func PrepareRepository(
+	ctx context.Context,
+	lib *siva.Library,
+	locID borges.LocationID,
+	repoID borges.RepositoryID,
+	endpoint string,
+	tmp billy.Filesystem,
+	clonePath string,
+) (borges.Repository, error) {
+	var r borges.Repository
 
-	for len(current.ParentHashes) > 0 {
-		current, err = current.Parent(0)
+	loc, err := lib.AddLocation(locID)
+	if err != nil {
+		if !siva.ErrLocationExists.Is(err) {
+			return nil, err
+		}
+
+		loc, err = lib.Location(locID)
+		if err != nil {
+			return nil, err
+		}
+
+		r, err = loc.Get(repoID, borges.RWMode)
+		if err != nil {
+			r, err = loc.Init(repoID)
+			if err != nil {
+				return nil, err
+			}
+		}
+	}
+
+	if r == nil {
+		r, err = createRootedRepo(ctx, loc, repoID, tmp, clonePath)
 		if err != nil {
 			return nil, err
 		}
 	}
 
-	return current, nil
+	if _, err := createRemote(r.R(), repoID.String(), endpoint); err != nil {
+		if cErr := r.Close(); cErr != nil {
+			err = fmt.Errorf("%s: %s", err.Error(), cErr.Error())
+		}
+
+		return nil, err
+	}
+
+	return r, nil
+}
+
+// FetchChanges fetches changes for the given remote into the borges.Repository.
+func FetchChanges(
+	ctx context.Context,
+	r borges.Repository,
+	remote string,
+	token string,
+) error {
+	opts := &git.FetchOptions{
+		RemoteName: remote,
+	}
+
+	if token != "" {
+		opts.Auth = &http.BasicAuth{
+			Username: "gitcollector",
+			Password: token,
+		}
+	}
+
+	if err := r.R().FetchContext(
+		ctx, opts,
+	); err != nil && err != git.NoErrAlreadyUpToDate {
+		if cErr := r.Close(); cErr != nil {
+			err = fmt.Errorf("%s: %s", err.Error(), cErr.Error())
+		}
+
+		return err
+	}
+
+	return nil
+}
+
+func createRootedRepo(
+	ctx context.Context,
+	loc borges.Location,
+	repoID borges.RepositoryID,
+	clonedFS billy.Filesystem,
+	clonedPath string,
+) (borges.Repository, error) {
+	repo, err := loc.Init(repoID)
+	if err != nil {
+		return nil, err
+	}
+
+	done := make(chan struct{})
+	go func() {
+		err = recursiveCopy(
+			"/", repo.FS(),
+			clonedPath, clonedFS,
+		)
+
+		close(done)
+	}()
+
+	select {
+	case <-done:
+	case <-ctx.Done():
+		err = ctx.Err()
+		repo.Close()
+		repo = nil
+	}
+
+	return repo, err
+}
+
+func recursiveCopy(
+	dst string,
+	dstFS billy.Filesystem,
+	src string,
+	srcFS billy.Filesystem,
+) error {
+	stat, err := srcFS.Stat(src)
+	if err != nil {
+		return err
+	}
+
+	if stat.IsDir() {
+		err = dstFS.MkdirAll(dst, stat.Mode())
+		if err != nil {
+			return err
+		}
+
+		files, err := srcFS.ReadDir(src)
+		if err != nil {
+			return err
+		}
+
+		for _, file := range files {
+			srcPath := filepath.Join(src, file.Name())
+			dstPath := filepath.Join(dst, file.Name())
+
+			err = recursiveCopy(dstPath, dstFS, srcPath, srcFS)
+			if err != nil {
+				return err
+			}
+		}
+	} else {
+		err = copyFile(dst, dstFS, src, srcFS, stat.Mode())
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func copyFile(
+	dst string,
+	dstFS billy.Filesystem,
+	src string,
+	srcFS billy.Filesystem,
+	mode os.FileMode,
+) error {
+	_, err := srcFS.Stat(src)
+	if err != nil {
+		return err
+	}
+
+	fo, err := srcFS.Open(src)
+	if err != nil {
+		return err
+	}
+	defer fo.Close()
+
+	fd, err := dstFS.OpenFile(dst, os.O_CREATE|os.O_TRUNC|os.O_WRONLY, mode)
+	if err != nil {
+		return err
+	}
+	defer fd.Close()
+
+	_, err = io.Copy(fd, fo)
+	if err != nil {
+		fd.Close()
+		dstFS.Remove(dst)
+		return err
+	}
+
+	return nil
 }
